Update app.txt
Ultra Sun/Ultra Moon wasn't localized. Also, the official name in Spanish has no space between Ultra and Sun/Moon, btw. Dunno why, it's on the official website. Looks kinda bad, but I went with it as it was on the official Spanish Pokémon website. Feel free to change it, of course.

Cheers!