Add save_path script method (#1145)
* Add save_path script method

Add a method that gets the path to the save file on the SD card.  This is needed if writing content outside of the save that references the save file.

* Add mkdir method to unistd

While this isn't the header where mkdir is usually found, adding all of sys/stat.h seems excessive, and unistd.h seems like a good fit since that's where rmdir is.  This also means that mode_t is defined in unistd as well.

* Clean up TitleLoader::savePath logic

Co-authored-by: Chris Feger <piepie62@users.noreply.github.com>